Make OpenMP REQUIRED, move check
OpenMP is now REQUIRED as the build will fail if it's not available.
In the process of making this change, I moved the check for OpenMP
from the src/ and tests/ dirs to the project root CMakeLists.txt,
which finds the package and creates the necessary targets for older
CMake versions (3.9+ will do it automatically).

The target is then made an INTERFACE link library of the openshot
target in src/CMakeLists.txt, which causes the OpenMP configs to
automatically propagate to the executables in both src/ and tests/.

The detection comes verbatim from this post, which also credits
Modern CMake: https://iscinumpy.gitlab.io/post/omp-on-high-sierra/